Fixed issue with multiplying Blotch
Also put in an automated effect change to see how it looks. My main complaint is that each change seems to cause a jarring jump in the animation. Perhaps there's some way to reuse the seed from the previous config? For now, I think I'll just try a different animation.